Merged: changes from linuxport. revs: 18781-18814
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/unified-pvr@18862 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/xbmc/GUIWindowVideoBase.cpp
@@ -908,7 +908,7 @@ bool CGUIWindowVideoBase::OnClick(int iItem)
 {
   CFileItemPtr pItem = m_vecItems->Get(iItem);
 
-  if (!pItem->m_bIsFolder && !pItem->Exists())
+  if (!pItem->m_bIsFolder && pItem->m_strPath.Find("playlist://") == CStdString::npos && !pItem->Exists())
   {
     if (!CGUIWindowVideoNav::DeleteItem(pItem.get(),true))
       return true;
